corrected showform and show-gif-form bugs
